requirements.txt
requests==2.31.0
apache-airflow-providers-github==2.5.1
PyGithub==2.3.0